Make category, language and skill name options appear on all three views
