--- src/__tests__/components/BotBuilder/BotBuilderPages/BuildViews/ConversationView.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import ConversationView from '../../../../../components/BotBuilder/BotBuilderPages/BuildViews/ConversationView';
-import { shallow } from 'enzyme';
-
-describe('<ConversationView />', () => {
-  it('render ConversationView without crashing', () => {
-    shallow(<ConversationView />);
-  });
-});
